Fixed incorrect file path when deleting data sources (#73)

+++ Daf.Meta.Editor/ViewModels/MainViewModel.cs
@@ -424,14 +424,20 @@ namespace Daf.Meta.Editor.ViewModels
 
 				if (result == MessageBoxResult.Yes)
 				{
-					// TODO: This points to the wrong path. Can not currently delete DataSources.
-					string dataSourcePath = Path.Combine(MetadataPath, "DataSources", $"{selectedDataSource.Name}.json");
+					string dataSourcePath = Path.Combine(MetadataPath, "DataSources", selectedDataSource.SourceSystem.ShortName, selectedDataSource.Tenant.ShortName, $"{selectedDataSource.Name}.json");
 
-					Model.RemoveDataSource(selectedDataSource.DataSource);
+					if (File.Exists(dataSourcePath))
+					{
+						Model.RemoveDataSource(selectedDataSource.DataSource);
 
-					DataSources.Remove(selectedDataSource);
+						DataSources.Remove(selectedDataSource);
 
-					File.Delete(dataSourcePath);
+						File.Delete(dataSourcePath);
+					}
+					else
+					{
+						MessageBox.Show($"The specified file path does not exist. Data Source could not be deleted.\n\nFile path: {dataSourcePath}", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+					}
 				}
 			}
 		}
